Beginnings of support for pseudomethods.
So far, all this does is handle desugaring of pseudomethod calls into
one-argument functions.  Started sketching out how to desugar into
multi-argument functions in the comments; that's not done yet.

The '::' syntax was suggested by pcwalton.  If we want to use ordinary
method call syntax (dot), then this becomes rather harder, because we
have to treat the pseudomethod call as a field expression and then
expect arbitrary expressions as the base of that field expression,
which breaks a lot of existing places in the compiler.